# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LemGendary Environment Manager API",
    "description": "REST and WebSocket sidecar service for LemGendary AI Studio.",
    "version": "2.0.0"
  },
  "paths": {
    "/api/health": {
      "get": {
        "summary": "Get Health",
        "description": "Retrieve complete ecosystem health audit.",
        "operationId": "get_health_api_health_get",
        "parameters": [
          {
            "name": "include_safety",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "If true, run per-project pip dry-runs to classify which outdated packages can be safely upgraded. Adds 15-40s of latency per Python project. Default false for fast polling.",
              "default": false,
              "title": "Include Safety"
            },
            "description": "If true, run per-project pip dry-runs to classify which outdated packages can be safely upgraded. Adds 15-40s of latency per Python project. Default false for fast polling."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/drift": {
      "get": {
        "summary": "Get Drift",
        "description": "Retrieve Python and NPM package drift data, manifest coverage, and\nsingle-manifest inventory.\n\nReturns:\n\n- ``python``: cross-project drift entries (packages in >= 2 manifests).\n- ``coverage``: per-project declared-vs-installed reconciliation.\n- ``single_manifest``: packages declared in exactly one manifest.\n- ``npm``: npm workspace drift entries.",
        "operationId": "get_drift_api_drift_get",
        "parameters": [
          {
            "name": "include_safety",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Same semantics as /api/health?include_safety.",
              "default": false,
              "title": "Include Safety"
            },
            "description": "Same semantics as /api/health?include_safety."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/hardware": {
      "get": {
        "summary": "Get Hardware",
        "operationId": "get_hardware_api_hardware_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/projects": {
      "get": {
        "summary": "Get Projects",
        "operationId": "get_projects_api_projects_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/npm": {
      "get": {
        "summary": "Get Npm",
        "operationId": "get_npm_api_npm_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/pipeline/status": {
      "get": {
        "summary": "Get Pipeline Status",
        "operationId": "get_pipeline_status_api_pipeline_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/pipeline/run": {
      "post": {
        "summary": "Run Pipeline",
        "operationId": "run_pipeline_api_pipeline_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunPipelineRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/update": {
      "post": {
        "summary": "Run Update",
        "description": "Trigger safe bottom-up package upgrades across all projects.",
        "operationId": "run_update_api_update_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/UpdateRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/manifests": {
      "get": {
        "summary": "Get Manifests",
        "operationId": "get_manifests_api_manifests_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/manifests/sync": {
      "post": {
        "summary": "Sync Manifests",
        "operationId": "sync_manifests_api_manifests_sync_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/clean": {
      "post": {
        "summary": "Clean Artifacts",
        "operationId": "clean_artifacts_api_clean_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/CleanRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/validate": {
      "post": {
        "summary": "Validate Codebase",
        "operationId": "validate_codebase_api_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ValidateRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/gui/state": {
      "get": {
        "summary": "Get Gui State",
        "description": "Aggregated ecosystem and host state for LemGendary AI Studio Desktop GUI.\n\nReturns host telemetry, hardware profile, discovered projects summary,\nand orchestrator status in a single non-blocking payload.",
        "operationId": "get_gui_state_api_gui_state_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/gui/ecosystem": {
      "get": {
        "summary": "Get Gui Ecosystem",
        "description": "Multi-sidecar health overview for LemGendary AI Studio Desktop GUI top-bar status.",
        "operationId": "get_gui_ecosystem_api_gui_ecosystem_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CleanRequest": {
        "properties": {
          "project": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project"
          }
        },
        "type": "object",
        "title": "CleanRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "RunPipelineRequest": {
        "properties": {
          "target_project": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Project"
          }
        },
        "type": "object",
        "title": "RunPipelineRequest"
      },
      "UpdateRequest": {
        "properties": {
          "project": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project"
          },
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run",
            "default": false
          }
        },
        "type": "object",
        "title": "UpdateRequest"
      },
      "ValidateRequest": {
        "properties": {
          "project": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project"
          }
        },
        "type": "object",
        "title": "ValidateRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}